Refactor filename for better reuse
