+++ src/types/character.ts
